[Silver I] Title: 쉬운 계단 수, Time: 64 ms, Memory: 11432 KB -BaekjoonHub

+++ 백준/Silver/10844. 쉬운 계단 수/쉬운 계단 수.java
@@ -24,12 +24,12 @@ public class Main {
             dp[i][0] = dp[i-1][1];
             dp[i][9] = dp[i-1][8];
             for (int j = 1; j < 9; j++) {
-                dp[i][j] = (int) ((long) dp[i-1][j-1] + dp[i-1][j+1]) % 1_000_000_000;
+                dp[i][j] = (dp[i-1][j-1] + dp[i-1][j+1]) % 1_000_000_000;
             }
         }
 
         for (int i = 1; i < 10; i++) {
-            answer = (int) ((long) answer + dp[n-1][i]) % 1_000_000_000;
+            answer = (answer + dp[n-1][i]) % 1_000_000_000;
         }
 
         System.out.println(answer);
